pyproject.toml
[tool.poetry]
name = "capstone-group8-backend"
authors = ["Daniel Leal <dlleal@uc.cl>"]
version = "0.0.1"
description = "Backend application for the group 8's capstone project."
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
google-api-python-client = "^2.2.0"
djangorestframework-simplejwt = "^4.6.0"
django-cors-headers = "^3.7.0"
exponent-server-sdk = "^2.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.4"
black = "^20.8b1"
flake8 = "^3.9.0"
isort = "^5.8.0"
pylint-django = "^2.4.3"
